backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9

